Reduce config structure: array is optional Migrate generation logic from duplicate files
